+++ src/router.tsx
@@ -19,8 +19,8 @@ import MyPage from '@/pages/MyPage';
 import MyPageEditPage from '@/pages/MyPageEditPage';
 import AccountManagePage from '@/pages/AccountManagePage';
 import LeavePage from '@/pages/LeavePage';
-import HouseWorkStepOnePage from '@/pages/HouseWorkStepOnePage';
-import HouseWorkStepTwoPage from '@/pages/HouseWorkStepTwoPage';
+import HouseWorkStepOnePage from '@/pages/housework/HouseWorkStepOnePage';
+import HouseWorkStepTwoPage from '@/pages/housework/HouseWorkStepTwoPage';
 import GroupInviteReceivePage from '@/pages/GroupInviteReceivePage';
 import ScrollToTop from '@/components/common/scroll/ScrollToTop';
 import { Toaster } from '@/components/common/ui/toaster';
